Enhance pallet management and user context handling
- Added `InOutRegister` constant for pallet operations.
- Introduced `GetByPalletAndCellAsync` in `PalletInCellRepository`.
- Changed `GetByIdAsync` signature in repositories to `object`.
- Enhanced task retrieval methods in `TaskRepository` with related entities.
- Implemented `GetTaskWithDetailsAsync` to fetch task details.
- Registered `ICurrentUserContext` service in `App.xaml.cs`.
- Updated `LoginWindow` to bind to `LoginViewModel`.
- Improved `BottomNavigationBarControl` with user profile and logout.
- Enhanced `RetrievalTaskItemControl` layout and bindings.
- Added properties in `LoginViewModel` for login status management.
- Introduced `SelectHndDestinationDialog` for HND retrieval selection.
- Enhanced `MamanHttpService` for fetching tasks with user context.
- Updated database context and migrations for task location fields.
- Added enums for OPC states to manage pallet movements.
- Refactored view models and services for better organization.
- Improved error handling and logging across the application.

+++ EM.Maman.DriverClient/ViewModels/MainViewModel.Initialization.cs
@@ -51,7 +51,7 @@ namespace EM.Maman.DriverClient.ViewModels
             {
                 using (var unitOfWork = _unitOfWorkFactory.CreateUnitOfWork())
                 {
-                    var dbTrolley = await unitOfWork.Trolleys.GetByIdAsync(activeTrolleyId.Value);
+                    var dbTrolley = await unitOfWork.Trolleys.GetByIdAsync((long)activeTrolleyId.Value); // Cast to long
 
                     if (dbTrolley == null)
                     {
